Add files via upload
wrote out the html file. included header that shows an h1 and h2, as well as display time limit of 60 seconds. body broken down into sections for starting the quizz, where the questions populate, the active timer and time left, submission of initials after quiz is done, and button to start over.